fhir-meds/pyproject.toml
[tool.poetry]
name = "fhirmeds"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fhirmeds", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fhirclient = "^4.1.0"
mysql-connector-python = "^8.1.0"
faker = "^19.11.0"
flask = "^2.0.0"
flask-login = "^0.5.0"
werkzeug = "2.0.0"
pyexecjs = "^1.5.1"
bcrypt = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
